Allow appending to sets before/after values
Sometimes when using a set the order of the values will be significant.
In these situations it is helpful to inject a new value before or after
another value.

For example, when writing a sequence of actions where the output of one
action is consumed by another action through composition.