Add sorted dictionary structure
There are some situations where an sorted hash is required.
For example when working with values that also have a quality or count,
it is often necessary to sort them before output.

To support this data type some functionality from Dictionary has been
split into HashStorage, similar to ValueStorage for lists, and a new
SortedStorage trait is added to be used instead of Storage.